+++ lib/runners/schema/trace.rb
@@ -12,7 +12,7 @@ module Runners
       )
       let :header, object(trace: literal(:header), message: string, recorded_at: string)
       let :warning, object(trace: literal(:warning), message: string, file: string?, recorded_at: string)
-      let :ci_config, object(trace: literal(:ci_config), content: any, file: string, recorded_at: string)
+      let :ci_config, object(trace: literal(:ci_config), content: any, raw_content: string, file: string, recorded_at: string)
       let :error, object(trace: literal(:error), message: string, recorded_at: string, truncated: boolean)
       let :anything, enum(command_line, status, stdout, stderr, message, header, warning, ci_config, error)
     end